pass through sort state to the header
