+++ client/reader/tag-stream/header.jsx
@@ -1,11 +1,18 @@
 import classnames from 'classnames';
 import { localize, translate } from 'i18n-calypso';
+import page from 'page';
 import PropTypes from 'prop-types';
 import { Component } from 'react';
 import { connect } from 'react-redux';
 import FollowButton from 'calypso/blocks/follow-button/button';
+import SegmentedControl from 'calypso/components/segmented-control';
+import { addQueryArgs } from 'calypso/lib/url';
 import ReaderFollowFeedIcon from 'calypso/reader/components/icons/follow-feed-icon';
 import ReaderFollowingFeedIcon from 'calypso/reader/components/icons/following-feed-icon';
+import { recordAction } from 'calypso/reader/stats';
+
+const updateQueryArg = ( params ) =>
+	page.replace( addQueryArgs( params, window.location.pathname + window.location.search ) );
 
 class TagStreamHeader extends Component {
 	static propTypes = {
@@ -14,11 +21,43 @@ class TagStreamHeader extends Component {
 		following: PropTypes.bool,
 		onFollowToggle: PropTypes.func,
 		showBack: PropTypes.bool,
+		showSort: PropTypes.bool,
+		sort: PropTypes.string,
 	};
 
+	useRelevanceSort = () => {
+		const sort = 'relevance';
+		recordAction( 'tag_page_clicked_relevance_sort' );
+		this.props.recordReaderTracksEvent &&
+			this.props.recordReaderTracksEvent( 'calypso_reader_clicked_tag_sort', {
+				tag: this.props.encodedTagSlug,
+				sort,
+			} );
+		updateQueryArg( { sort } );
+	};
+
+	useDateSort = () => {
+		const sort = 'date';
+		recordAction( 'tag_page_clicked_date_sort' );
+		this.props.recordReaderTracksEvent &&
+			this.props.recordReaderTracksEvent( 'calypso_reader_clicked_tag_sort', {
+				tag: this.props.encodedTagSlug,
+				sort,
+			} );
+		updateQueryArg( { sort } );
+	};
 	render() {
-		const { title, description, isPlaceholder, showFollow, following, onFollowToggle, showBack } =
-			this.props;
+		const {
+			title,
+			description,
+			isPlaceholder,
+			showFollow,
+			following,
+			onFollowToggle,
+			showBack,
+			showSort,
+		} = this.props;
+		const sortOrder = this.props.sort || 'recent';
 
 		// A bit of a hack: check for a prompt tag (which always have a description) from the slug before waiting for tag info to load,
 		// so we can set a smaller title size and prevent it from resizing as the page loads. Should be refactored if tag descriptions
@@ -37,6 +76,21 @@ class TagStreamHeader extends Component {
 				<div className="tag-stream__header-title-group">
 					<h1 className="tag-stream__header-title">{ title }</h1>
 					{ description && <h2 className="tag-stream__header-description">{ description }</h2> }
+				</div>
+				<div className="tag-stream__header-sort-picker">
+					{ showSort && (
+						<SegmentedControl compact>
+							<SegmentedControl.Item
+								selected={ sortOrder !== 'date' }
+								onClick={ this.useRelevanceSort }
+							>
+								{ this.props.translate( 'Popular' ) }
+							</SegmentedControl.Item>
+							<SegmentedControl.Item selected={ sortOrder === 'date' } onClick={ this.useDateSort }>
+								{ this.props.translate( 'Recent' ) }
+							</SegmentedControl.Item>
+						</SegmentedControl>
+					) }
 				</div>
 				<div className="tag-stream__header-follow">
 					{ showFollow && (
