+++ client/reader/tag-stream/main.jsx
@@ -1,16 +1,13 @@
 import { getUrlParts } from '@automattic/calypso-url';
 import { localize } from 'i18n-calypso';
 import { find } from 'lodash';
-import page from 'page';
 import PropTypes from 'prop-types';
 import { Component } from 'react';
 import { connect } from 'react-redux';
 import QueryReaderFollowedTags from 'calypso/components/data/query-reader-followed-tags';
 import QueryReaderTag from 'calypso/components/data/query-reader-tag';
-import SegmentedControl from 'calypso/components/segmented-control';
 import { navigate } from 'calypso/lib/navigate';
 import { createAccountUrl } from 'calypso/lib/paths';
-import { addQueryArgs } from 'calypso/lib/url';
 import ReaderMain from 'calypso/reader/components/reader-main';
 import HeaderBack from 'calypso/reader/header-back';
 import { recordAction, recordGaEvent } from 'calypso/reader/stats';
@@ -25,14 +22,10 @@ import EmptyContent from './empty';
 import TagStreamHeader from './header';
 import './style.scss';
 
-const updateQueryArg = ( params ) =>
-	page.replace( addQueryArgs( params, window.location.pathname + window.location.search ) );
-
 class TagStream extends Component {
 	static propTypes = {
 		encodedTagSlug: PropTypes.string,
 		decodedTagSlug: PropTypes.string,
-		sort: PropTypes.string,
 	};
 
 	state = {
@@ -104,37 +97,11 @@ class TagStream extends Component {
 		);
 	};
 
-	useRelevanceSort = () => {
-		const sort = 'relevance';
-		recordAction( 'tag_page_clicked_relevance_sort' );
-		this.props.recordReaderTracksEvent( 'calypso_reader_clicked_tag_sort', {
-			tag: this.props.encodedTagSlug,
-			sort,
-		} );
-		updateQueryArg( { sort } );
-	};
-
-	useDateSort = () => {
-		const sort = 'date';
-		recordAction( 'tag_page_clicked_date_sort' );
-		this.props.recordReaderTracksEvent( 'calypso_reader_clicked_tag_sort', {
-			tag: this.props.encodedTagSlug,
-			sort,
-		} );
-		updateQueryArg( { sort } );
-	};
-
 	render() {
 		const emptyContent = () => <EmptyContent decodedTagSlug={ this.props.decodedTagSlug } />;
 		const title = this.props.decodedTagSlug;
 		const tag = find( this.props.tags, { slug: this.props.encodedTagSlug } );
-		const sortOrder = this.props.sort || 'relevance';
 		const titleText = title.replace( /-/g, ' ' );
-
-		// const wideDisplay = this.props.width > WIDE_DISPLAY_CUTOFF;
-		// const segmentedControlClass = wideDisplay
-		// 	? 'search-stream__sort-picker is-wide'
-		// 	: 'search-stream__sort-picker';
 
 		let imageSearchString = this.props.encodedTagSlug;
 
@@ -155,6 +122,7 @@ class TagStream extends Component {
 						title={ title }
 						imageSearchString={ imageSearchString }
 						showFollow={ false }
+						showSort={ false }
 						showBack={ this.props.showBack }
 					/>
 					{ emptyContent() }
@@ -173,19 +141,10 @@ class TagStream extends Component {
 					following={ this.isSubscribed() }
 					onFollowToggle={ this.toggleFollowing }
 					showBack={ this.props.showBack }
+					showSort={ true }
+					sort={ this.props.sort }
+					recordReaderTracksEvent={ this.props.recordReaderTracksEvent }
 				/>
-
-				<SegmentedControl compact className="tag-stream__sort-picker">
-					<SegmentedControl.Item
-						selected={ sortOrder !== 'date' }
-						onClick={ this.useRelevanceSort }
-					>
-						{ this.props.translate( 'Relevance' ) }
-					</SegmentedControl.Item>
-					<SegmentedControl.Item selected={ sortOrder === 'date' } onClick={ this.useDateSort }>
-						{ this.props.translate( 'Date' ) }
-					</SegmentedControl.Item>
-				</SegmentedControl>
 			</>
 		);
 
@@ -214,13 +173,14 @@ class TagStream extends Component {
 }
 
 export default connect(
-	( state, { decodedTagSlug } ) => {
+	( state, { decodedTagSlug, sort } ) => {
 		const tag = getReaderTagBySlug( state, decodedTagSlug );
 		return {
 			description: tag?.description,
 			followedTags: getReaderFollowedTags( state ),
 			tags: getReaderTags( state ),
 			isLoggedIn: isUserLoggedIn( state ),
+			sort,
 		};
 	},
 	{
